+++ packages/grafana-ui/src/components/ThresholdsEditor/_ThresholdsEditor.scss
