+++ packages/grafana-ui/src/components/index.scss
@@ -1,4 +1,4 @@
 @import 'CustomScrollbar/CustomScrollbar';
 @import 'DeleteButton/DeleteButton';
-@import 'Thresholds/Thresholds';
+@import 'ThresholdsEditor/ThresholdsEditor';
 @import 'Tooltip/Tooltip';
