+++ packages/grafana-ui/src/components/index.ts
@@ -5,4 +5,4 @@ export { CustomScrollbar } from './CustomScrollbar/CustomScrollbar';
 export { ColorPicker } from './ColorPicker/ColorPicker';
 export { SeriesColorPickerPopover } from './ColorPicker/SeriesColorPickerPopover';
 export { SeriesColorPicker } from './ColorPicker/SeriesColorPicker';
-export { ThresholdsEditor } from './Thresholds/Thresholds';
+export { ThresholdsEditor } from './ThresholdsEditor/ThresholdsEditor';
